Make instructions more usefully non-specific
Setting a definite (and pretty tight) time limit sends the wrong
message, especially since we haven't done much measurement of how long
this actually tends to take people. Instead, let's just express that we
don't want it to be a burden, but let people judge for themselves
exactly how many hours they want or need to spend on it.